feat(push-notifications): having push notifications as wrapper inside app
