+++ App.tsx
@@ -4,6 +4,7 @@ import { Box, NativeBaseProvider } from 'native-base';
 import useCachedResources from './src/hooks/useCachedResources';
 import useColorScheme from './src/hooks/useColorScheme';
 import Navigation from './src/navigation';
+import PushNotificationWrapper from './src/wrappers/PushNotificationWrapper';
 
 export default function App() {
   const isLoadingComplete = useCachedResources();
@@ -18,11 +19,12 @@ export default function App() {
   } else {
     return (
       <NativeBaseProvider>
-
+        <PushNotificationWrapper />
         <Navigation colorScheme={colorScheme} />
         <StatusBar />
         <Box>Hello world</Box>
         </NativeBaseProvider>
+
 
     );
   }
